% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{datareceipt_has_key}
\alias{datareceipt_has_key}
\title{Is an API key available?}
\usage{
datareceipt_has_key()
}
\value{
\code{TRUE} if a key is set in the \code{DATARECEIPT_API_KEY} environment
variable, \code{FALSE} otherwise.
}
\description{
Is an API key available?
}
\examples{
datareceipt_has_key()
}
